# Optimized the adaptive_avg_pool2d op when output_size == 1