… paddle/fluid/operators/pool_op.cu