<!-- Demo: https://github.com/PaddlePaddle/Paddle/pull/24810 -->
### PR types
Performance optimization
### PR changes
<!-- One of [ OPs | APIs | Docs | Others ] -->
OPs 
### Describe
<!-- Describe what this PR does -->
Optimized the adaptive_avg_pool2d op when output_size == 1


<!--Test result--->
case 规模{4, 2048, 64, 128},output_size{1,1}，NCHW

  |   |   |   |   |  
-- | -- | -- | -- | -- | --
paddle reduce | 优化前： old | 优化后： new | pytorch | 性能提升：(old-new)/new
float16 | 5.7325ms | 156.43us | 157.63us | 35.64倍 |  
float32 | 5.7778ms | 307.55us | 309.78us | 17.78倍 |  
float64 | 5.8074ms | 606.06us | 611.50us | 8.58倍 |  