@zhangting2020 2021-02-24
PR描述里用表格列出优化前后的kernel耗时，以及提升幅度。